making library, cleaning up directory
