--- MCU/checkThreshold.cpp
@@ -1,14 +1,0 @@
-﻿/*
- * checkThreshold.cpp
- *
- * Created: 1/30/2018 10:29:34 PM
- * Author : Pol Sieira
- */ 
-
-//#include "sam.h"
-#include "checkThreshold.h"
-
-int checkThreshold()
-{
-	return 0; /*TODO: Change to actual threshold checking method*/
-}
