--- MCU/checkThreshold.h
@@ -1,16 +1,0 @@
-/*
- * checkThreshold.cpp
- *
- * Created: 1/30/2018 10:29:34 PM
- * Author : Pol Sieira
- */ 
-
-#ifndef CHECKTHRESHOLD_H_
-#define CHECKTHRESHOLD_H_
-
-//#include "sam.h"
-#include "checkThreshold.h"
-
-int checkThreshold();
-
-#endif
