--- MCU/faultCheckFS.h
@@ -1,16 +1,0 @@
-/*
- * faultCheckFS.h
- *
- * Created: 1/29/2018 3:29:24 PM
- * Author : Pol Sieira
- */ 
-
-#ifndef FAULTCHECKFS_H_
-#define FAULTCHECKFS_H_
-
-//#include "sam.h"
-#include "faultCheckFS.h"
-
-int faultCheckFS();
-
-#endif
