--- MCU/faultCheckRW.h
@@ -1,16 +1,0 @@
-/*
- * faultCheckRW.h
- *
- * Created: 1/29/2018 3:29:24 PM
- * Author : Pol Sieira
- */ 
-
-#ifndef FAULTCHECKRW_H_
-#define FAULTCHECKRW_H_
-
-//#include "sam.h"
-#include "faultCheckRW.h"
-
-int faultCheckRW();
-
-#endif
