+++ MCU/faultManagement.cpp
@@ -7,9 +7,6 @@
 
 //#include "sam.h"
 #include "faultManagement.h"
-#include "faultCheckRW.h"
-#include "faultCheckFS.h"
-#include "recovery.h"
 
 void faultManagement(){
 	
@@ -52,7 +49,6 @@ void faultManagement(){
 				}
 				else
 				{
-					/*TODO: Function to Shut off power to RW*/
 					return;
 				}
 			}
@@ -70,4 +66,70 @@ void faultManagement(){
 			}
 		}	
 	}
+}
+
+int checkThreshold()
+{
+	return 0; /*TODO: Change to actual threshold checking method*/
+}
+
+int faultCheck()
+{
+	/*Defining the Variables*/
+	int faultDetected, faultTimerActive;
+	
+	/*Run threshold check*/
+	faultDetected = checkThreshold();
+	
+	/*Is a fault detected?*/
+	if (faultDetected == 1)
+	{
+		if (faultTimerActive == 1)
+		{
+			/*TODO: if (time < 30)
+			{
+				return;
+			}
+			else
+			{
+				timer_delete();
+				faultDetected = 1;
+			}
+			*/
+		}
+		else
+		{
+			/*TODO: timer_create*/
+		}
+	}
+	else
+	{
+		if (faultTimerActive == 1)
+		{
+			/*TODO: timer_delete*/	
+		}
+		else
+		{
+			return;
+		}
+	}
+	return 0;
+}
+
+
+void recovery(int faultType)
+{
+	if (faultType == 0)
+	{
+		throw;
+	} 
+	if (faultType == 1)
+	{
+		/*TODO: Switch command to RW2*/
+	}
+	if (faultType == 2)
+	{
+		/*TODO: Switch sensor reading to FS2*/
+	}
+return;
 }
