+++ MCU/faultManagement.h
@@ -4,17 +4,12 @@
  * Created: 1/29/2018 2:41:34 PM
  * Author : Pol Sieira
  */ 
-
 #ifndef FAULTMANAGEMENT_H_
 #define FAULTMANAGEMENT_H_
 
-//#include "sam.h"
-#include <math.h>
-#include "faultManagement.h"
-#include "faultCheckRW.h"
-#include "faultCheckFS.h"
-#include "recovery.h"
-
 void faultManagement();
-
+int checkThreshold();
+int faultCheckFS();
+int faultCheckRW();
+void recovery(int faultType);
 #endif
